+++ app/src/androidTest/java/com/github/se/studybuddies/screens/CreateGroupScreen.kt
@@ -22,9 +22,8 @@ class CreateGroupScreen(semanticsProvider: SemanticsNodeInteractionsProvider) :
   val saveButton: KNode = createGroupCol.child { hasTestTag("save_button") }
   val profileButton: KNode = createGroupCol.child { hasTestTag("set_picture_button") }
 
-
   val saveButtonText: KNode = onNode { hasTestTag("save_button_text") }
-  /*    
+  /*
       semanticsProvider = semanticsProvider, viewBuilderAction = { hasTestTag("create_group") }) {
   val content: KNode = onNode { hasTestTag("content") }
   val column: KNode = content.child { hasTestTag("CreateGroup") }*/
