+++ app/src/androidTest/java/com/github/se/studybuddies/tests/AccountSettingsTest.kt
@@ -1,33 +1,20 @@
 package com.github.se.studybuddies.tests
 
-import android.net.Uri
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.se.studybuddies.data.User
 import com.github.se.studybuddies.navigation.NavigationActions
 import com.github.se.studybuddies.navigation.Route
 import com.github.se.studybuddies.screens.AccountSettingsScreen
-import com.github.se.studybuddies.ui.settings.AccountSettings
-import com.github.se.studybuddies.viewModels.UserViewModel
-import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
-import io.github.kakaocup.compose.node.element.ComposeScreen
-import io.mockk.confirmVerified
-import io.mockk.impl.annotations.RelaxedMockK
-import io.mockk.junit4.MockKRule
-import io.mockk.verify
-
-import androidx.compose.ui.test.junit4.createComposeRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.se.studybuddies.navigation.NavigationActions
-import com.github.se.studybuddies.navigation.Route
 import com.github.se.studybuddies.ui.settings.AccountSettings
 import com.github.se.studybuddies.viewModels.UserViewModel
 import com.kaspersky.components.composesupport.config.withComposeSupport
 import com.kaspersky.kaspresso.kaspresso.Kaspresso
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
 import io.github.kakaocup.compose.node.element.ComposeScreen
+import io.mockk.confirmVerified
 import io.mockk.impl.annotations.RelaxedMockK
 import io.mockk.junit4.MockKRule
+import io.mockk.verify
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
 import org.junit.Before
@@ -61,6 +48,7 @@ class AccountSettingsTest : TestCase(kaspressoBuilder = Kaspresso.Builder.withCo
           signOutButton { assertIsDisplayed() }
         }
   }
+
   @Test
   fun topAppBar() = run {
     ComposeScreen.onComposeScreen<AccountSettingsScreen>(composeTestRule) {
@@ -104,7 +92,6 @@ class AccountSettingsTest : TestCase(kaspressoBuilder = Kaspresso.Builder.withCo
         }
     // verify { mockNavActions.navigateTo(Route.LOGIN) }
     // confirmVerified(mockNavActions)
-    
+
   }
-  
 }
