+++ app/src/androidTest/java/com/github/se/studybuddies/tests/CreateGroupTest.kt
@@ -2,6 +2,7 @@ package com.github.se.studybuddies.tests
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.test.espresso.action.ViewActions.closeSoftKeyboard
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.se.studybuddies.navigation.NavigationActions
 import com.github.se.studybuddies.navigation.Route
@@ -11,12 +12,15 @@ import com.github.se.studybuddies.viewModels.GroupViewModel
 import com.kaspersky.components.composesupport.config.withComposeSupport
 import com.kaspersky.kaspresso.kaspresso.Kaspresso
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
+import io.github.kakaocup.compose.node.element.ComposeScreen
 import io.github.kakaocup.compose.node.element.ComposeScreen.Companion.onComposeScreen
 import io.mockk.Called
 import io.mockk.confirmVerified
 import io.mockk.impl.annotations.RelaxedMockK
 import io.mockk.junit4.MockKRule
 import io.mockk.verify
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.runBlocking
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -40,36 +44,23 @@ class CreateGroupTest : TestCase(kaspressoBuilder = Kaspresso.Builder.withCompos
   }
 
   @Test
-  fun nameGroupInput() = run {
-    onComposeScreen<CreateGroupScreen>(composeTestRule) {
-      step("Check group name") {
-        groupField {
-          // arrange: verify pre-conditions
-          assertIsDisplayed()
-
-          // act: interact with the text field
-          performClick()
-
-          // assert: check that both the label and placeholder are correct
-          assertTextContains("Group Name")
-          assertTextContains("Enter a group name")
+  fun inputGroupName() {
+    ComposeScreen.onComposeScreen<com.github.se.studybuddies.screens.CreateGroupScreen>(
+        composeTestRule) {
+          saveButton { assertIsNotEnabled() }
+          groupField {
+            performTextClearance()
+            performTextInput("Official Group Testing")
+            assertTextContains("Official Group Testing")
+          }
+          closeSoftKeyboard()
+          saveButton {
+            assertIsEnabled()
+            performClick()
+          }
         }
-      }
-
-      step("Enter group name") {
-        // arrange: verify pre-conditions + enter search query
-        groupField {
-          performTextClearance()
-          performTextInput("Test name")
-        }
-
-        // act: click on Nominatim's proposition
-        groupFieldProposal { performClick() }
-
-        // assert: check the suggestion box proposition
-        groupField { assertTextContains(value = "Test name", substring = true) }
-      }
-    }
+    verify { mockNavActions.navigateTo(Route.GROUPSHOME) }
+    confirmVerified(mockNavActions)
   }
 
   @Test
@@ -78,14 +69,20 @@ class CreateGroupTest : TestCase(kaspressoBuilder = Kaspresso.Builder.withCompos
       step("Open group screen") {
         groupField {
           assertIsDisplayed()
-
           // interact with the text field
           performClick()
-
           // clear the text field
           performTextClearance()
         }
+        closeSoftKeyboard()
+        saveButton {
+          // arrange: verify pre-conditions
+          assertIsDisplayed()
+          assertIsNotEnabled()
 
+          // act: click on the save button
+          performClick()
+        }
         /*
         saveButtonText {
           assertIsDisplayed()
@@ -98,6 +95,25 @@ class CreateGroupTest : TestCase(kaspressoBuilder = Kaspresso.Builder.withCompos
         confirmVerified(mockNavActions)
       }
     }
+  }
+
+  fun elementsAreDisplayed() {
+    ComposeScreen.onComposeScreen<com.github.se.studybuddies.screens.CreateGroupScreen>(
+        composeTestRule) {
+          runBlocking {
+            delay(6000) // Adjust the delay time as needed
+          }
+          // content { assertIsDisplayed() }
+          // groupNameField { assertIsDisplayed() }
+          profileButton {
+            assertIsDisplayed()
+            assertHasClickAction()
+          }
+          saveButton {
+            assertIsDisplayed()
+            assertHasClickAction()
+          }
+        }
   }
 
   @Test
